// host-api/node/src/public/gateway/hosts/dto/create-host.dto.ts
import { CreateHostInterface, CustomHeader, Domain } from "@app/common/interface";

export class CreateHostDto implements CreateHostInterface {
    domains: Domain[];
    custom_headers: CustomHeader[];
    destination: string;
    domain_name: string;
    path: string;
    has_custom_headers: boolean;
    websocket_support: boolean;
    force_https_redirect: boolean;
    allow_caching: boolean;
    block_exploits: boolean;
    auto_ssl: boolean;
    publicly_accessible: boolean;
      // ip_whitelist: false,
            // allowed_ips: string[]
}